requirements.txt
asn1crypto==1.2.0
certifi==2019.9.11
cffi==1.13.2
chardet==3.0.4
click==6.7
configparser==3.5.0
cryptography==2.8
cycler==0.10.0
cymem==1.31.2
cytoolz==0.9.0.1
datasketch==1.2.7
decorator==4.3.0
dill==0.2.8.2
docutils==0.14
elasticsearch==6.3.1
Flask==1.0.2
Flask-Cors==3.0.6
graphviz==0.9
idna==2.7
itsdangerous==0.24
Jinja2==2.10.1
kiwisolver==1.0.1
lxml==4.2.5
MarkupSafe==1.0
matplotlib==3.0.0
mkl-fft==1.0.6
mkl-random==1.0.1
msgpack==0.5.6
msgpack-numpy==0.4.4.1
murmurhash==0.28.0
networkx==2.2
nltk==3.4.5
numpy==1.15.2
pandas==0.25.1
pbr==4.3.0
plac==0.9.6
preshed==1.0.1
py-eureka-client==0.7.4
pybliometrics==2.2.0
pycparser==2.19
pyOpenSSL==19.0.0
pyparsing==2.2.2
pyreadline==2.1
PySocks==1.7.1
pytextrank==1.1.0
python-dateutil==2.7.3
pytz==2019.2
redis==2.10.6
regex==2018.7.11
requests==2.21.0
scipy==1.1.0
six==1.11.0
spacy==2.0.12
statistics==1.0.3.5
TBB==0.1
termcolor==1.1.0
thinc==6.10.3
toolz==0.9.0
tornado==6.0.3
tqdm==4.26.0
ujson==1.35
urllib3==1.24.3
Werkzeug==0.14.1
win-inet-pton==1.1.0
wincertstore==0.2
wrapt==1.10.11
